00_Docs/Prerequisites/The_AI_Engineer_Onramp_Cohort_2/Session_02_Back_End_Web_App_Development_&_Deployment_of_LLM_Applications/sample_backend_scripts/pyproject.toml
[project]
name = "sample-backend-scripts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pypdf2>=3.0.1",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "openai>=1.53.0",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.9",
    "pydantic>=2.0.0",
]
